backend/requirements.txt
PyPDF2
textract
python-docx
